Modifying the Server File into a Module
We modify the server file into a module to allow the code to be resusable and at
the same time avoid any naming conflicts that might arise by polluting the
global namespace. Since the CommonJS module system separates modules into their
own contexts, any interference between modules won't occur.

Additionally, we intend to separate functionality into modules to improve
readability, maintainability and scalability. For instance, we will need to
route requests to the correct handler, a function that handles and processes a
response. So we'll need two modules: a router module and a handler module. But
this is a discussion for a later time.

As of now we use index.js to start up our server and house any specifics to our
application.

index.js:
        - starts up the server and houses any specifics to our server (objects,
          functions, etc.)

        - we allow node to use this as an entry point into the code (for java
          users: think of it as a driver class with a main function, a main file).

server.js:
        - this file is now a module that is exported as an object. The object
          exposes the start functionality as method.