Adding the Router and Integrating Routing in the Server
How do we separate the functionality of the server with the functionality of the
application -- functionality that is based on the url, the url's parameters, and
the HTTP request methods (GET/POST)?

Instead of having the server do all this work, we can create a router to route
the program, on the event of a request, to a specific handler based on whatever
parameters we choose.  Additionally, we can also pass request objects if we ever
need to pass it to the handler.

The router module, right now, only contains a function. But the idea is to
create a module that can route the program, on the event of a request, to a
specific handler based on the url and the accompanying HTTP request methods.

We also perform a dependency injection in passing the route function as an
argument to the server. This allows us to lightly couple the server and the
router. By allowing the server to generalize the route function -- by invoking
the function that was passed as an argument--, we can avoid creating a
dependency on a specific router and its routing method (in other words, we avoid
specifically naming a router in the server). As a result, we allow the server to
be coupled with any router, not just our own.

server.js:
        - Generalizes routing to a route function. We do not import the router
          as it creates an unnecssary dependency on a router. Instead, we accept
          a route function that the server can invoke to being the routing process. As a
          result, we allow the server to be coupled with any router, not just our own.

router.js:
        - Adding the file

        - Contains a basic function that prints the path name

        - This file will have the ability to invoke specific functions based on
          the url, url parameters, and the HTTP request method (get/post)

index.js:
        - Since this file's purpose is to run the application and is specific to
          the application, then we create dependencies here. We import the
          router and pass the route method into the server.