Adding Handlers File and Implementing Routing
We implement minimal routing functionality into route using the path name and a
map of path names to their handlers. The route method is passed a object from the
server containing a pair of pathnames and their respective handlers. We match
the given path name to an existing entry in the map and invoke the corresponding
function.

The handle map is housed in the index file because creating it anywhere else
will break the reusability of the other modules. Index.js is specific to our
application, so it makes sense to house objects specific to our application
here.

index.js:
        - A map of pathnames to handlers is created in this file instead of in
          the router. Creating the map in the router would destroy its
          reusability since we're hard-coding routes specific to our application, routes
          that won't be applicable to other applications. So, we pass this map
          over to the server which, in turn, will pass it over to the route method.

server.js:
        - Accepts the handle map and passes it over to the routing method.

        - As a sidenote, if you are wondering how we can pass an object over to
          a function that returns immediately and still retain access to it in
          future requests, the answer is that the callback onRequest, which is a closure,
          will carries with it the environment (outer variables, etc.) it was created in.
          Therefore, an invocation of onRequest, by createServer, will always be able to
          access the handle object even after start returns.

router.js:
        - route uses the handle map to invoke the function that corresponds to
          the given pathname.

handlers.js:
        - The bare mimimum is being added. At the moment, the handlers can only print to the
          console when invoked by the route method.