Allowing the Handlers to Serve Responses
We finally reached a point where we allow handlers to serve responses to
requests. Thanks to our routing mechanism, this simply becomes a matter of
"bringing the server to the content" instead of the other way around. Bringing
content to the server module requires us to pass back data and it assumes that
the route or the handler that is invoked by route is synchronous. Unfortunately,
we can guarantee this. If an alternative, async route method is passed in, or if
the handler is async or uses solely async functions, then you can expect route()
to return immediately without any content. The result would be passing back null
content to the server module, which would eventually get written as a response.
Recall that async functions does not force us to wait for their results.
Instead, the function returns immediately while it performs the computation, and
then it invokes the callback -- a function we defined on the invocation of the
async function -- with its result passed into it. It is in this function that we
make us of the results. So you see we can't rely on synchronous patterns when
asynchronous functions are used.

So instead of passing back the handlers data, we pass the response object over
to the handlers for them to use, in whatever capacity, to form the response. This
allows the response object to be referenced in callbacks if the handlers decide
to use async functions.

handlers.js:
        - There are two handlers so far: start and upload. The handler object
          maps '/start' to the start handler and '/upload' to the upload
          handler.

        - The handlers actually do stuff now! The start handler writes "Hello
          World" as a response. The upload handler writes "UPLOAD!".

router.js:
        - It passes the response object as an argument to the handler.

server.js:
        - It passes the response object to the route method so that it can pass
          it on to the handler.