Handling POST Requests and Their Data
When a POST request is recieved, the data associated with it is accessible
through the request object, though not always a trivial process. The process in
which we read post data is listening for "data" events emitted from the request
object when a data chunk is available, and then invoking the appropriate
callback to process that data. The callback is to accept the chunks of data read
from the request as an argument. Since the data read in chunks, we continuously
collect and join the chunks into a single variable outside of the callback.
When the request has no more data available for consumption, it emits an "end"
event indicating that the stream has completed. The callback for this event is
an appropriate place to route since the data has been collected and is ready to be
handed off to the handler.

(As a side note, an end event is always emitted from a request object regardless
of its http method. This allows us to place the route invocation within the
"end" callback and still have it be invoked on a get request [such as a request
for "/start"].)

The request objects is more than just an object, up until this point, the most
we have done is access its properties. The request object s capable more as it
implements both the EventEmitter and ReadableStream interfaces. As both an
EventEmitter and a ReadableStream, it emits a "data" event when a chunk of data is
ready to be consumed. Additionally, it also emits an "end" event when the stream
has not more data to be consumed. Therefore, we need listeners to register the
appropriate callback, an event handler, for the particular event and to listen
for such events.

The responsibility of reading and collecting the data is being placed on the server.
It seems reasonable at this time to have the server extract the post data seeing
as we also extract header information here. However, this may change in the
future.

Lastly, the start handler now serves an html form that is used by the
browser to create an http post request. The form consists of a text box and a
submit button. The form instructs the browser to create a request directed to
'/upload', meaning that the upload handler will respond to this request.

handlers.js:
        - The start handler now responds with an html form consisting of a text
          box and submit button. The form instructs the browser to send the
          contents in the form of a POST request with the url '/upload'.

              - Now, normally we would separate the presentation from the logic,
                but I don't know how to add a templating engine to this yet. So,
                this will do for now.

                        - I suppose I could read in the html content from the
                          file system, but that seems excessive for such a small
                          form.

        - At the moment, the upload takes the data given to it by the server,
          via the route, and we simply return it as a response to show that the
          text was sent successfully.

router.js:
        - The route method now accepts post data to be passed in so that it can
          be passed on to the appropriate handler.

server.js:
        - Event listeners are placed on the request object for when it emits a
          "data" event or "end" event.

        - The "data" handler, the callback registered with the "data" listener,
          joins the chunks of data into a single outer variable.

                - As a side note, the callback, a closure, carries with it the
                  environment (outer variables, etc.) it was defined in because it
                  references an outer variable like postData. postData will then
                  persists even after the context it was created in, onRequest, completes.

        - When the "end" event is emitted, the router can finally route the
          program to the appropriate handler with the data extracted from the
          request.

                - As a side note, an end event is always emitted from a request
                  object regardless of its http method. This allows us to place
                  the route invocation within the "end" callback and still have
                  it be invoked on a get request (such as a request for "/start").