Parsing the Query String
As we have seen, the body of a post request contains the client's data. In this
case, our data is formatted in the form of a query string. We parse the
query string with the help of the querystring module; then, we extract the text by
referencing the field of the parsed string.

handlers.js:

        - the querystring module parses the query string and we extract the text
          sent in the text input box under the field "text". Note that input
          field's label is "text", thus the field where our text resides is
          "text".

        - upload now displays the uploaded format in plain text