Prelude to Uploading Images
Recall that the initial idea of this application was to allow the user to upload
an image and be able to serve the uploaded image back to the user. As a prelude
to implementing the upload mechanism, we tackle the issue of serving content
stored on our local file system.

Thankfully, node provides us with the ability to read from the hard drive using
the 'fs' module. The image is read, as binary data, from '/tmp' and piped into
the response object. Finally, the response object sends the response to the
client, satisfying our objective of serving the image to the user.

We also make sure to expose the show handler (using exports.show) and map it to
'/show'.

As a side note: "Pipe" is a term specific to streams or processes. The act of passing
data from one stream (or process) to another is referred to as a "pipe".

handlers.js:
        - the show handler reads an image from the hard drive, using the fs
          module, and writes the image to the response.
                - As a side note: this is an instance where the being able to
                  write to the response stream in parts is extremely useful. The
                  image is most likely to be read in pieces, therefore we want to be
                  able to pipe it into the response stream without waiting to
                  piece together the entire image.

                - In the event of an error, the error message is sent as a response.

index.js:
        - "/show" is mapped to the show handler