Implementing the Upload Functionality
The upload functionality can be broken down into two parts:
    - adding an upload element on the /start page
    - reading and storing the files onto the server

We need the ability to upload a file from the start page, which makes for a
great starting point. An enctype attribute with a value of 'multipart/form-data'
is added to the form action element. The default 'x-www-urlencoded' which
specifies a url encoding of the body, while suitable for text forms, is
ineffective for file types. If a form contains an <input type='file'>, then we
need to use an enctype of 'multipart/form-data' for the form action. In our
form, we allow for only one file to be uploaded, nothing else.

The shift from uploading text to uploading files requires that we alter our code
to accommodate this change. The code was initially written to handle a url
encoding of the body; with the change in the encoding type, the way the request
and its data is handle also has to change.

First, we want the upload handler to handle upload requests. Currently, requests
are handled by the server module. There is no specific mechanism to dictate for
which handlers a request gets passed, so, requests will have to be passed onto
all handlers.  In doing so, we modify functions in the routing module
and the handler module to take a request parameter, and we also have to modify
the arguments passed into invocations of route() and handler(). Note that we
allow some of the handlers to omit the request parameter because they aren't
required -- functions invoked with more arguments than parameters discard the
superfluous arguments.

Since we're removing the request handling from the server, there is no need for
data to be consumed in the server code. Additionally, this code conflicts with
our use of the node-formidable module. The module consumes data events -- also
the data associated with these events -- from the request object. Data can only
be consumed once, node.js doesn't buffer data, therefore keeping both pieces of
code will ensure that by the time node-formidable has a chance to consume the
data in the request, it will have already been consumed. So all the data
listeners and end listeners are removed.

The node-formidable module allows us to abstract away the reading, storing, and
parsing of data from the request body. To do this, it uses the metaphor of a
form being submitted via HTTP POST. To realize this abstraction, we use the
constructor IncomingForm() to create an new instance. The form then can be used
to parse our request object into files and fields using the parse() method.
parse() requires a request and a callback -- with the the parameters error,
fields, and files -- as arguments.

When the form successfully parses the request, asynchronously, a response is
formed: an internal error message if the parse fails or, if an image is
uploaded, an html page with the uploaded imaged embedded in it. Also, the
uploaded file is renamed to the hard-coded filename in /show because the show
handler needs to be able to find and read the image in order to respond to any
requests for it. More importantly, the html page that results from an upload
request uses /show to request the uploaded image in order to display it.

package.json:
    - Introduces formidable as a dependency

handlers.js:
    - Changing the html form in the /start handler to allow the client to submit
      files as part of a request.
        - The form uses an encoding type "multipart/form-data" which allows for
          files to be split into parts. Note: multiple requests aren't issued in
          order to send a file in parts. The data is sent as chunks as part
          of one request. This is how files are transmitted using html. (See MDN
          article on Transfer-Encoding.)

          The default is "x-www-urlencoded" which formats data as a query string
          and sends it as part of the body in the request, however this isn't
          suitable for files, only for text fields.

        - The request doesn't have the data immediately, we would have to read
          it in chunks. However, we are using formidable, which does all the
          work for us.

    - We include the module formidable as dependency. formidable has the
      functionality to parse a request object into files and fields. It reads
      files and store them, in addition to reading fields, included in request
      (a single file can require a client to send multiple chunks).

        - Formidable reads the data, parses it, and stores the files onto
          the server. It uses the metaphor of a form to abstract away
          the details of parsing incoming files.

            - It does so asynchronously, so we have to pass a callback when
              parsing a request.

        - Files are stored in /tmp by default

    - /upload handler is responsible for handling the request and forming a response.

    - In our case, we rename the file to the hardcoded filename in /show. The
      file show.png needs to exist in /tmp if we are to read it and serve it as
      a response.

        - Note: It's possible that the rename won't work until we unlink the
          filename from the file currently holding that filename.

    - As a response, we embed and display the uploaded image as an html page.
      The client uses the /show resource to grab the image from the server.

        - show reads the uploaded file into memory and is ready to embed it
          into an html page.

server.js:
    - Removing 'data' and 'end' listeners since we're placing the responsibility
      of parsing and extracting data from the request on the handlers
      instead of the server

router.js:
    - Renamed the parameter to reflect the type of objects being passed in